% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stud_dates_ca1009}
\alias{stud_dates_ca1009}
\title{dates of 'Published' field for 1009 cancer studies (some were NA in study_publ_dates, remedied)}
\format{named vector}
\usage{
stud_dates_ca1009
}
\description{
dates of 'Published' field for 1009 cancer studies (some were NA in study_publ_dates, remedied)
}
\keyword{datasets}
